Refactor deploy for use outside of cli
